+++ backend/src/controllers/user.controller.ts
@@ -49,7 +49,20 @@ class userController {
 
   async googleSignIn(req: Request, res: Response) {}
 
-  async kakaoSignIn(req: Request, res: Response) {}
+  // 'kakao'
+  async kakaoSignIn(req: Request, res: Response) {
+    console.log('[Kakao] kakao login');
+    passport.authenticate('login-kakao');
+  }
+
+  // 'kakao/callback'
+  async kakaoCallback(req: Request, res: Response, next: NextFunction) {
+    console.log('[Kakao] kakao callback');
+    passport.authenticate('login-kakao', {
+      successRedirect: '/',
+      failureRedirect: '/', // kakaoStrategy에서 실패한다면 실행
+    });
+  }
 
   getUserData(req: Request, res: Response) {}
