+++ backend/src/routes/user.router.ts
@@ -7,4 +7,8 @@ userRouter.get('/', userController.temp);
 
 userRouter.post('/login', userController.signIn);
 
+userRouter.get('/kakao', userController.kakaoSignIn);
+
+userRouter.get('/kakao/callback', userController.kakaoCallback);
+
 export default userRouter;
